fix: resolve CI failures for Hytte-aut7n
Move the wardrobe API helpers (ApiError, api, messageFor) out of
WardrobePage.tsx into pages/wardrobeApi.ts — exporting non-components
from a component file trips react-refresh/only-export-components.
